+++ src/patch/core/Global.zig
@@ -1,5 +1,8 @@
 const Self = @This();
 
+const GlobalState = @import("SharedDef.zig").GlobalState;
+const GlobalFunction = @import("SharedDef.zig").GlobalFunction;
+
 const std = @import("std");
 const win = std.os.windows;
 
@@ -15,6 +18,10 @@ const dbg = @import("../util/debug.zig");
 const msg = @import("../util/message.zig");
 const mem = @import("../util/memory.zig");
 
+const app = @import("../appinfo.zig");
+const VERSION = app.VERSION;
+const VERSION_STR = app.VERSION_STR;
+
 const rti = @import("racer").Time;
 const rg = @import("racer").Global;
 const rrd = @import("racer").RaceData;
@@ -26,153 +33,74 @@ const w32 = @import("zigwin32");
 const w32kb = w32.ui.input.keyboard_and_mouse;
 const w32xc = w32.ui.input.xbox_controller;
 const w32wm = w32.ui.windows_and_messaging;
+const POINT = w32.foundation.POINT;
 const KS_DOWN: i16 = -1;
 const KS_PRESSED: i16 = 1; // since last call
 
 // NOTE: may want to figure out all the code caves in .data for potential use
 // TODO: split up the versioning, global structs, etc. from the business logic
 
-// VERSION
-
-// NOTE: current minver for updates: 0.1.2
-// TODO: include tag when appropriate
-pub const Version = std.SemanticVersion{
-    .major = 0,
-    .minor = 1,
-    .patch = 5,
-    //.pre = "alpha",
-    .build = "373",
-};
-
-// TODO: use SemanticVersion parse fn instead
-pub const VersionStr: [:0]u8 = s: {
-    var buf: [127:0]u8 = undefined;
-    break :s std.fmt.bufPrintZ(&buf, "Annodue {d}.{d}.{d}.{s}", .{
-        Version.major,
-        Version.minor,
-        Version.patch,
-        Version.build.?,
-    }) catch unreachable; // comptime
-};
-
-pub const PLUGIN_VERSION = 19;
-
 // STATE
 
-const RaceState = enum(u8) { None, PreRace, Countdown, Racing, PostRace, PostRaceExiting };
+fn global_player_reset(self: *GlobalState) void {
+    const p = &self.player;
+    p.upgrades_lv = rrd.PLAYER.*.pFile.upgrade_lv; // TODO: remove from gs, now that it's easy?
+    p.upgrades_hp = rrd.PLAYER.*.pFile.upgrade_hp; // TODO: remove from gs, now that it's easy?
+    p.upgrades = for (0..7) |i| {
+        if (p.upgrades_lv[i] > 0 and p.upgrades_hp[i] > 0) break true;
+    } else false;
 
-const GLOBAL_STATE_VERSION = 5;
+    p.flags1 = 0;
+    p.boosting = .Off;
+    p.underheating = .On; // you start the race underheating
+    p.overheating = .Off;
+    p.dead = .Off;
+    p.deaths = 0;
 
-// TODO: move all references to patch_memory to use internal allocator; add
-// allocator interface to GlobalFunction
-// TODO: move all the common game check stuff from plugins/modules to here; cleanup
-// TODO: add index of currently consumed loaded tga IDs, since they are arbitrarily assigned
-//   also, some kind of interface plugins can use to avoid clashes
-//   list of stuff to update when it's made:
-//     inputdisplay, practice mode vis, spare camstates used
-pub const GlobalState = extern struct {
-    patch_memory: [*]u8 = undefined,
-    patch_size: usize = undefined,
-    patch_offset: usize = undefined,
+    p.heat_rate = re.Test.PLAYER.*.stats.HeatRate; // TODO: remove from gs, now that it's easy?
+    p.cool_rate = re.Test.PLAYER.*.stats.CoolRate; // TODO: remove from gs, now that it's easy?
+    p.heat = 0;
+}
 
-    init_late_passed: bool = false,
+fn global_player_update(self: *GlobalState) void {
+    const p = &self.player;
+    p.flags1 = re.Test.PLAYER.*.flags1; // TODO: remove from gs, now that it's easy?
+    p.heat = re.Test.PLAYER.*.temperature; // TODO: remove from gs, now that it's easy?
+    const engine = re.Test.PLAYER.*.engineStatus; // TODO: remove from gs, now that it's easy?
 
-    practice_mode: bool = false,
-
-    hwnd: ?win.HWND = null,
-    hinstance: ?win.HINSTANCE = null,
-
-    dt_f: f32 = 0,
-    fps: f32 = 0,
-    fps_avg: f32 = 0,
-    timestamp: u32 = 0,
-    framecount: u32 = 0,
-
-    in_race: st.ActiveState = .Off,
-    race_state: RaceState = .None,
-    race_state_prev: RaceState = .None,
-    race_state_new: bool = false,
-    player: extern struct {
-        upgrades: bool = false,
-        upgrades_lv: [7]u8 = undefined,
-        upgrades_hp: [7]u8 = undefined,
-
-        flags1: u32 = 0,
-        boosting: st.ActiveState = .Off,
-        underheating: st.ActiveState = .On,
-        overheating: st.ActiveState = .Off,
-        dead: st.ActiveState = .Off,
-        deaths: u32 = 0,
-
-        heat_rate: f32 = 0,
-        cool_rate: f32 = 0,
-        heat: f32 = 0,
-    } = .{},
-
-    fn player_reset(self: *GlobalState) void {
-        const p = &self.player;
-        p.upgrades_lv = rrd.PLAYER.*.pFile.upgrade_lv; // TODO: remove from gs, now that it's easy?
-        p.upgrades_hp = rrd.PLAYER.*.pFile.upgrade_hp; // TODO: remove from gs, now that it's easy?
-        p.upgrades = for (0..7) |i| {
-            if (p.upgrades_lv[i] > 0 and p.upgrades_hp[i] > 0) break true;
-        } else false;
-
-        p.flags1 = 0;
-        p.boosting = .Off;
-        p.underheating = .On; // you start the race underheating
-        p.overheating = .Off;
-        p.dead = .Off;
-        p.deaths = 0;
-
-        p.heat_rate = re.Test.PLAYER.*.stats.HeatRate; // TODO: remove from gs, now that it's easy?
-        p.cool_rate = re.Test.PLAYER.*.stats.CoolRate; // TODO: remove from gs, now that it's easy?
-        p.heat = 0;
-    }
-
-    fn player_update(self: *GlobalState) void {
-        const p = &self.player;
-        p.flags1 = re.Test.PLAYER.*.flags1; // TODO: remove from gs, now that it's easy?
-        p.heat = re.Test.PLAYER.*.temperature; // TODO: remove from gs, now that it's easy?
-        const engine = re.Test.PLAYER.*.engineStatus; // TODO: remove from gs, now that it's easy?
-
-        p.boosting.update((p.flags1 & (1 << 23)) > 0);
-        p.underheating.update(p.heat >= 100);
-        p.overheating.update(for (0..6) |i| {
-            if (engine[i] & (1 << 3) > 0) break true;
-        } else false);
-        p.dead.update((p.flags1 & (1 << 14)) > 0);
-        if (p.dead == .JustOn) p.deaths += 1;
-    }
-};
+    p.boosting.update((p.flags1 & (1 << 23)) > 0);
+    p.underheating.update(p.heat >= 100);
+    p.overheating.update(for (0..6) |i| {
+        if (engine[i] & (1 << 3) > 0) break true;
+    } else false);
+    p.dead.update((p.flags1 & (1 << 14)) > 0);
+    if (p.dead == .JustOn) p.deaths += 1;
+}
 
 pub var GLOBAL_STATE: GlobalState = .{};
 
-pub const GLOBAL_FUNCTION_VERSION = 15;
-
-pub const GlobalFunction = extern struct {
+pub var GLOBAL_FUNCTION: GlobalFunction = .{
     // Settings
-    SettingGetB: *const @TypeOf(settings.get_bool) = &settings.get_bool,
-    SettingGetI: *const @TypeOf(settings.get_i32) = &settings.get_i32,
-    SettingGetU: *const @TypeOf(settings.get_u32) = &settings.get_u32,
-    SettingGetF: *const @TypeOf(settings.get_f32) = &settings.get_f32,
+    .SettingGetB = &settings.get_bool,
+    .SettingGetI = &settings.get_i32,
+    .SettingGetU = &settings.get_u32,
+    .SettingGetF = &settings.get_f32,
     // Input
-    InputGetKb: *const @TypeOf(input.get_kb) = &input.get_kb,
-    InputGetKbRaw: *const @TypeOf(input.get_kb_raw) = &input.get_kb_raw,
-    InputGetMouse: *const @TypeOf(input.get_mouse_raw) = &input.get_mouse_raw,
-    InputGetMouseDelta: *const @TypeOf(input.get_mouse_raw_d) = &input.get_mouse_raw_d,
-    InputLockMouse: *const @TypeOf(input.lock_mouse) = &input.lock_mouse,
-    //InputGetMouseInWindow: *const @TypeOf(input.get_mouse_inside) = &input.get_mouse_inside,
-    InputGetXInputButton: *const @TypeOf(input.get_xinput_button) = &input.get_xinput_button,
-    InputGetXInputAxis: *const @TypeOf(input.get_xinput_axis) = &input.get_xinput_axis,
+    .InputGetKb = &input.get_kb,
+    .InputGetKbRaw = &input.get_kb_raw,
+    .InputGetMouse = &input.get_mouse_raw,
+    .InputGetMouseDelta = &input.get_mouse_raw_d,
+    .InputLockMouse = &input.lock_mouse,
+    //InputGetMouseInWindow= &input.get_mouse_inside,
+    .InputGetXInputButton = &input.get_xinput_button,
+    .InputGetXInputAxis = &input.get_xinput_axis,
     // Game
-    GameFreezeEnable: *const @TypeOf(freeze.Freeze.freeze) = &freeze.Freeze.freeze,
-    GameFreezeDisable: *const @TypeOf(freeze.Freeze.unfreeze) = &freeze.Freeze.unfreeze,
-    GameFreezeIsFrozen: *const @TypeOf(freeze.Freeze.is_frozen) = &freeze.Freeze.is_frozen,
+    .GameFreezeEnable = &freeze.Freeze.freeze,
+    .GameFreezeDisable = &freeze.Freeze.unfreeze,
+    .GameFreezeIsFrozen = &freeze.Freeze.is_frozen,
     // Toast
-    ToastNew: *const @TypeOf(toast.ToastSystem.NewToast) = &toast.ToastSystem.NewToast,
+    .ToastNew = &toast.ToastSystem.NewToast,
 };
-
-pub var GLOBAL_FUNCTION: GlobalFunction = .{};
 
 // UTIL
 
@@ -185,7 +113,7 @@ fn DrawMenuPracticeModeLabel() void {
 }
 
 fn DrawVersionString() void {
-    rt.DrawText(36, 480 - 24, "{s}", .{VersionStr}, 0xFFFFFFFF, null) catch {};
+    rt.DrawText(36, 480 - 24, "{s}", .{VERSION_STR}, 0xFFFFFFFF, null) catch {};
 }
 
 // INIT
@@ -229,8 +157,8 @@ pub fn EngineUpdateStage14A(gs: *GlobalState, _: *GlobalFunction) callconv(.C) v
     };
     gs.race_state_new = gs.race_state != gs.race_state_prev;
 
-    if (gs.race_state_new and gs.race_state == .PreRace) gs.player_reset();
-    if (gs.in_race.on()) gs.player_update();
+    if (gs.race_state_new and gs.race_state == .PreRace) global_player_reset(gs);
+    if (gs.in_race.on()) global_player_update(gs);
 }
 
 pub fn TimerUpdateA(gs: *GlobalState, _: *GlobalFunction) callconv(.C) void {