+++ src/patch/dll_savestate.zig
@@ -5,9 +5,10 @@ const std = @import("std");
 const w32 = @import("zigwin32");
 const VIRTUAL_KEY = w32.ui.input.keyboard_and_mouse.VIRTUAL_KEY;
 
-const GlobalSt = @import("core/Global.zig").GlobalState;
-const GlobalFn = @import("core/Global.zig").GlobalFunction;
-const COMPATIBILITY_VERSION = @import("core/Global.zig").PLUGIN_VERSION;
+const GlobalSt = @import("appinfo.zig").GLOBAL_STATE;
+const GlobalFn = @import("appinfo.zig").GLOBAL_FUNCTION;
+const COMPATIBILITY_VERSION = @import("appinfo.zig").COMPATIBILITY_VERSION;
+const VERSION_STR = @import("appinfo.zig").VERSION_STR;
 
 const debug = @import("core/Debug.zig");
 
@@ -16,6 +17,8 @@ const st = @import("util/active_state.zig");
 const scroll = @import("util/scroll_control.zig");
 const msg = @import("util/message.zig");
 const mem = @import("util/memory.zig");
+const TemporalCompressor = @import("util/temporal_compression.zig").TemporalCompressor;
+const TDataPoint = @import("util/temporal_compression.zig").DataPoint;
 
 const rg = @import("racer").Global;
 const ri = @import("racer").Input;
@@ -80,48 +83,21 @@ const LoadState = enum(u32) {
     ScrubExiting,
 };
 
+const RewindDataType = TemporalCompressor(4, 4, 4);
+
 const state = struct {
     var savestate_enable: bool = false;
+    var initialized: bool = false;
 
     var rec_state: LoadState = .Recording;
-    var initialized: bool = false;
-    var frame: usize = 0;
-    var frame_total: usize = 0;
-    var last_framecount: u32 = 0;
-
-    const off_input: usize = 0;
-    const off_race: usize = ri.COMBINED_SIZE;
-    const off_test: usize = off_race + rrd.SIZE;
-    const off_hang: usize = off_test + re.Test.SIZE;
-    const off_cman: usize = off_hang + re.Hang.SIZE;
-    const off_END: usize = off_cman + re.cMan.SIZE;
-
-    const frames: usize = 60 * 60 * 8; // 8min @ 60fps
-    //const frame_size: usize = off_cman + rc.cMan.SIZE;
-    const header_size: usize = std.math.divCeil(usize, off_END, 4 * 8) catch unreachable; // comptime
-    const header_type: type = std.packed_int_array.PackedIntArray(u1, header_bits);
-    const header_bits: usize = off_END / 4;
-    const offsets_off: usize = 0;
-    const offsets_size: usize = frames * 4;
-    const headers_off: usize = offsets_off + offsets_size;
-    const headers_size: usize = header_size * frames;
-    const stage_off: usize = headers_off + headers_size;
-    const data_off: usize = stage_off + off_END * 2;
-
-    // FIXME: remove gpa/alloc, do some kind of core integration instead
-    var gpa: ?std.heap.GeneralPurposeAllocator(.{}) = null;
-    var alloc: ?std.mem.Allocator = null;
-    const memory_size: usize = 1024 * 1024 * 64; // 64MB
-    var memory: []u8 = undefined;
-    var memory_addr: usize = undefined;
-    var memory_end_addr: usize = undefined;
-    var raw_offsets: [*]u8 = undefined;
-    var raw_headers: [*]u8 = undefined;
-    var raw_stage: [*]u8 = undefined;
-    var offsets: *[frames]usize = undefined;
-    var headers: *[frames]header_type = undefined;
-    var stage: *[2][off_END / 4]u32 = undefined;
-    var data: [*]u8 = undefined;
+    var rec_data: RewindDataType = .{};
+    var rec_sources = [_]TDataPoint{
+        .{ .data = @as([*]u8, @ptrFromInt(ri.COMBINED_ADDR))[0..ri.COMBINED_SIZE] }, // Input
+        .{}, // RaceData
+        .{}, // Test
+        .{}, // Hang
+        .{}, // cMan
+    };
 
     var load_delay: usize = 500; // ms
     var load_time: usize = 0;
@@ -154,55 +130,13 @@ const state = struct {
         return scrub_input_inc.gets() == s;
     }
 
-    const layer_size: isize = 4;
-    const layer_depth: isize = 4;
-    var layer_widths: [layer_depth]usize = widths: {
-        var widths: [layer_depth]usize = undefined;
-        for (1..layer_depth + 1) |f| {
-            widths[layer_depth - f] = std.math.pow(usize, layer_size, f);
-        }
-        break :widths widths;
-    };
-    var layer_indexes: [layer_depth + 1]usize = undefined;
-    var layer_index_count: usize = undefined;
-
-    fn init(_: *GlobalFn) void {
-        if (initialized) return;
-        defer initialized = true;
-
-        gpa = std.heap.GeneralPurposeAllocator(.{}){};
-        alloc = gpa.?.allocator();
-        memory = alloc.?.alloc(u8, memory_size) catch @panic("failed to allocate memory for savestate/rewind");
-        @memset(memory[0..memory_size], 0x00);
-
-        memory_addr = @intFromPtr(memory.ptr);
-        memory_end_addr = @intFromPtr(memory.ptr) + memory_size;
-        raw_offsets = memory.ptr + offsets_off;
-        raw_headers = memory.ptr + headers_off;
-        raw_stage = memory.ptr + stage_off;
-        data = memory.ptr + data_off;
-        offsets = @as(@TypeOf(offsets), @ptrFromInt(memory_addr + offsets_off));
-        headers = @as(@TypeOf(headers), @ptrFromInt(memory_addr + headers_off));
-        stage = @as(@TypeOf(stage), @ptrFromInt(memory_addr + stage_off));
-    }
-
-    fn deinit(_: *GlobalFn) void {
-        if (!initialized) return;
-        defer initialized = false;
-
-        if (alloc) |a| a.free(memory);
-        if (gpa) |_| switch (gpa.?.deinit()) {
-            .leak => @panic("leak detected when deinitializing savestate/rewind"),
-            else => {},
-        };
-
-        reset();
-    }
-
     fn reset() void {
-        if (frame == 0) return;
-        frame = 0;
-        frame_total = 0;
+        rec_data.reset();
+        //rec_sources[0].data = @ptrFromInt(ri.COMBINED_ADDR); // don't need to reset this
+        rec_sources[1].data = rrd.PLAYER_SLICE.*;
+        rec_sources[2].data = re.Test.PLAYER_SLICE.*;
+        rec_sources[3].data = re.Manager.entitySlice(.Hang, 0);
+        rec_sources[4].data = re.Manager.entitySlice(.cMan, 0);
         load_frame = 0;
         load_count = 0;
         scrub_frame = 0;
@@ -212,9 +146,7 @@ const state = struct {
     // FIXME: better new-frame checking that doesn't only account for tabbing out
     // i.e. also when pausing, physics frozen with ingame feature, etc.
     fn saveable(gs: *GlobalSt) bool {
-        const space_ok: bool = memory_end_addr - @intFromPtr(data) - offsets[frame] >= off_END;
-        const frames_ok: bool = frame < frames - 1;
-        return gs.in_race.on() and space_ok and frames_ok;
+        return gs.in_race.on() and rec_data.canSave();
     }
 
     // FIXME: check if you're actually in the racing part, also integrate with global
@@ -240,110 +172,6 @@ const state = struct {
         return race_ok and !tabbed_out and !paused and loading_ok;
     }
 
-    fn get_depth(index: usize) usize {
-        var depth: usize = layer_depth;
-        var depth_test: usize = index;
-        while (depth_test % layer_size == 0 and depth > 0) : (depth -= 1) {
-            depth_test /= layer_size;
-        }
-        return depth;
-    }
-
-    fn set_layer_indexes(index: usize) void {
-        layer_index_count = 0;
-        var last_base: usize = 0;
-        for (layer_widths) |w| {
-            const remainder = index % w;
-            const base = index - remainder;
-            if (base > 0 and base != last_base) {
-                last_base = base;
-                layer_indexes[layer_index_count] = base;
-                layer_index_count += 1;
-            }
-            if (remainder < layer_size) {
-                if (remainder == 0) break;
-                layer_indexes[layer_index_count] = index;
-                layer_index_count += 1;
-                break;
-            }
-        }
-    }
-
-    fn uncompress_frame(index: usize, skip_last: bool) void {
-        @memcpy(raw_stage[0..off_END], data[0..off_END]);
-
-        set_layer_indexes(index);
-        var indexes: usize = layer_index_count - @intFromBool(skip_last);
-        if (indexes == 0) return;
-
-        for (layer_indexes[0..indexes]) |l| {
-            const header = &headers[l];
-            const frame_data = @as([*]usize, @ptrFromInt(@intFromPtr(data) + offsets[l]));
-            var j: usize = 0;
-            for (0..header_bits) |h| {
-                if (header.get(h) == 1) {
-                    stage[0][h] = frame_data[j];
-                    j += 1;
-                }
-            }
-        }
-    }
-
-    // FIXME: in future, probably can skip the first step each new frame, because
-    // the most recent frame would already be in stage1 from last time
-    fn save_compressed(gs: *GlobalSt, gv: *GlobalFn) void {
-        // FIXME: why the fk does this crash if it comes after the guard
-        if (!initialized) init(gv);
-        if (!saveable(gs)) return;
-        last_framecount = gs.framecount;
-
-        var data_size: usize = 0;
-        if (frame > 0) {
-            uncompress_frame(frame, true);
-
-            const s1_base = raw_stage + off_END;
-            mem.read_bytes(ri.COMBINED_ADDR, s1_base + off_input, ri.COMBINED_SIZE);
-            @memcpy(s1_base + off_race, rrd.PLAYER_SLICE.*);
-            @memcpy(s1_base + off_test, re.Test.PLAYER_SLICE.*);
-            @memcpy(s1_base + off_hang, re.Manager.entitySlice(.Hang, 0));
-            @memcpy(s1_base + off_cman, re.Manager.entitySlice(.cMan, 0));
-
-            var header = &headers[frame];
-            header.setAll(0);
-            var new_frame = @as([*]u32, @ptrFromInt(@intFromPtr(data) + offsets[frame]));
-            var j: usize = 0;
-            for (0..header_bits) |h| {
-                if (stage[0][h] != stage[1][h]) {
-                    header.set(h, 1);
-                    new_frame[j] = stage[1][h];
-                    data_size += 4;
-                    j += 1;
-                }
-            }
-        } else {
-            data_size = off_END;
-            mem.read_bytes(ri.COMBINED_ADDR, data + off_input, ri.COMBINED_SIZE);
-            @memcpy(data + off_race, rrd.PLAYER_SLICE.*);
-            @memcpy(data + off_test, re.Test.PLAYER_SLICE.*);
-            @memcpy(data + off_hang, re.Manager.entitySlice(.Hang, 0));
-            @memcpy(data + off_cman, re.Manager.entitySlice(.cMan, 0));
-        }
-        frame += 1;
-        offsets[frame] = offsets[frame - 1] + data_size;
-    }
-
-    fn load_compressed(index: usize, gs: *GlobalSt) void {
-        if (!loadable(gs)) return;
-
-        uncompress_frame(index, false);
-        _ = mem.write_bytes(ri.COMBINED_ADDR, &raw_stage[off_input], ri.COMBINED_SIZE);
-        @memcpy(rrd.PLAYER_SLICE.*, raw_stage[off_race..off_test]); // WARN: maybe perm issues
-        @memcpy(re.Test.PLAYER_SLICE.*, raw_stage[off_test..off_hang]); // WARN: maybe perm issues
-        @memcpy(re.Manager.entitySlice(.Hang, 0).ptr, raw_stage[off_hang..off_cman]);
-        @memcpy(re.Manager.entitySlice(.cMan, 0).ptr, raw_stage[off_cman..off_END]);
-        frame = index + 1;
-    }
-
     fn handle_settings(gf: *GlobalFn) callconv(.C) void {
         savestate_enable = gf.SettingGetB("savestate", "enable") orelse false;
         load_delay = gf.SettingGetU("savestate", "load_delay") orelse 500;
@@ -352,13 +180,14 @@ const state = struct {
 
 // LOADER LOGIC
 
-fn DoStateRecording(gs: *GlobalSt, gf: *GlobalFn) LoadState {
-    state.save_compressed(gs, gf);
+fn DoStateRecording(gs: *GlobalSt, _: *GlobalFn) LoadState {
+    if (state.saveable(gs))
+        state.rec_data.save(gs.framecount);
 
     if (state.save_input_st.gets() == .JustOn) {
-        state.load_frame = state.frame - 1;
+        state.load_frame = state.rec_data.frame - 1;
     }
-    if (state.save_input_ld.gets() == .JustOn and state.frames > 0) {
+    if (state.save_input_ld.gets() == .JustOn and state.rec_data.frames > 0) {
         state.load_time = state.load_delay + gs.timestamp;
         return .Loading;
     }
@@ -368,12 +197,13 @@ fn DoStateRecording(gs: *GlobalSt, gf: *GlobalFn) LoadState {
 
 fn DoStateLoading(gs: *GlobalSt, _: *GlobalFn) LoadState {
     if (state.save_input_ld.gets() == .JustOn) {
-        state.scrub_frame = std.math.cast(i32, state.frame).? - 1;
-        state.frame_total = state.frame;
+        state.scrub_frame = std.math.cast(i32, state.rec_data.frame).? - 1;
+        state.rec_data.frame_total = state.rec_data.frame;
         return .Scrubbing;
     }
     if (gs.timestamp >= state.load_time) {
-        state.load_compressed(state.load_frame, gs);
+        if (!state.loadable(gs)) return .Recording;
+        state.rec_data.restore(state.load_frame);
         state.load_count += 1;
         return .Recording;
     }
@@ -382,7 +212,7 @@ fn DoStateLoading(gs: *GlobalSt, _: *GlobalFn) LoadState {
 
 fn DoStateScrubbing(gs: *GlobalSt, _: *GlobalFn) LoadState {
     if (state.save_input_st.gets() == .JustOn) {
-        state.load_frame = state.frame - 1;
+        state.load_frame = state.rec_data.frame - 1;
     }
     if (state.save_input_ld.gets() == .JustOn) {
         state.load_frame = @min(state.load_frame, std.math.cast(u32, state.scrub_frame).?);
@@ -392,19 +222,21 @@ fn DoStateScrubbing(gs: *GlobalSt, _: *GlobalFn) LoadState {
 
     state.scrub_frame = state.scrub.UpdateEx(
         state.scrub_frame,
-        std.math.cast(i32, state.frame_total).?,
+        std.math.cast(i32, state.rec_data.frame_total).?,
         false,
     );
 
-    state.load_compressed(std.math.cast(u32, state.scrub_frame).?, gs);
+    if (!state.loadable(gs)) return .Scrubbing;
+    state.rec_data.restore(std.math.cast(u32, state.scrub_frame).?);
     return .Scrubbing;
 }
 
 fn DoStateScrubExiting(gs: *GlobalSt, _: *GlobalFn) LoadState {
-    state.load_compressed(std.math.cast(u32, state.scrub_frame).?, gs);
+    if (state.loadable(gs))
+        state.rec_data.restore(std.math.cast(u32, state.scrub_frame).?);
 
     if (state.save_input_st.gets() == .JustOn) {
-        state.load_frame = state.frame - 1;
+        state.load_frame = state.rec_data.frame - 1;
     }
 
     if (gs.timestamp < state.load_time) return .ScrubExiting;
@@ -416,8 +248,17 @@ fn DoStateScrubExiting(gs: *GlobalSt, _: *GlobalFn) LoadState {
 fn UpdateState(gs: *GlobalSt, gv: *GlobalFn) void {
     if (!state.updateable(gs)) return;
 
-    if (gs.race_state != .Racing)
-        return state.reset();
+    if (!state.initialized) {
+        defer state.initialized = true;
+        state.reset();
+        state.rec_data.sources = state.rec_sources[0..];
+        state.rec_data.init();
+    }
+
+    if (gs.race_state_new and gs.race_state == .PreRace)
+        state.reset();
+
+    if (gs.race_state != .Racing) return;
 
     state.rec_state = switch (state.rec_state) {
         .Recording => DoStateRecording(gs, gv),
@@ -447,8 +288,8 @@ export fn OnInit(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
 
 export fn OnInitLate(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
 
-export fn OnDeinit(_: *GlobalSt, gf: *GlobalFn) callconv(.C) void {
-    state.deinit(gf);
+export fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
+    state.rec_data.deinit();
 }
 
 // HOOKS
@@ -483,76 +324,9 @@ export fn EarlyEngineUpdateA(gs: *GlobalSt, _: *GlobalFn) callconv(.C) void {
     // TODO: experiment with conditionally showing each string; only show fr
     // if playing back, only show st if a frame is actually saved?
     if (gs.practice_mode and gs.race_state == .Racing) {
-        rt.DrawText(16, 480 - 16, "Fr {d}", .{state.frame}, null, null) catch {};
+        rt.DrawText(16, 480 - 16, "Fr {d}", .{state.rec_data.frame}, null, null) catch {};
         rt.DrawText(92, 480 - 16, "St {d}", .{state.load_frame}, null, null) catch {};
         if (state.load_count > 0)
             rt.DrawText(168, 480 - 16, "Ld {d}", .{state.load_count}, null, null) catch {};
     }
-}
-
-// COMPRESSION-RELATED FUNCTIONS
-// not really in use/needs work, basically just stuff for testing
-
-// TODO: move to compression lib whenever that happens
-
-// FIXME: assumes array of raw data; rework to adapt it to new compressed data
-fn save_file() void {
-    const file = std.fs.cwd().createFile("annodue/testdata.bin", .{}) catch |err| return msg.ErrMessage("create file", @errorName(err));
-    defer file.close();
-
-    const middle = state.frame * state.frame_size;
-    const end = state.frames * state.frame_size;
-    _ = file.write(state.data[middle..end]) catch return;
-    _ = file.write(state.data[0..middle]) catch return;
-}
-
-// FIXME: dumped from patch.zig; need to rework into a generalized function
-// TODO: cleanup unreachable after migrating to lib
-fn check_compression_potential() void {
-    const savestate_size: usize = 0x2428 / 4;
-    const savestate_count: usize = 128;
-    const savestate_head: usize = savestate_size / 8;
-    const layer_size: usize = 4;
-    const layer_depth: usize = 4;
-
-    const testfile = std.fs.cwd().openFile("annodue/testdata.bin", .{}) catch unreachable;
-    defer testfile.close();
-    const reportfile = std.fs.cwd().createFile("annodue/testreport.txt", .{}) catch unreachable;
-    defer reportfile.close();
-
-    var data = std.mem.zeroes([layer_depth + 1][savestate_size]u32);
-    var total_bytes: usize = 0;
-
-    for (0..savestate_count - 1) |i| {
-        var depth: usize = layer_depth;
-        var depth_test: usize = i;
-        while (depth_test % layer_size == 0 and depth > 0) {
-            depth_test /= layer_size;
-            depth -= 1;
-        }
-
-        _ = testfile.read(@as(*[savestate_size * 4]u8, @ptrCast(&data[depth]))) catch unreachable;
-        if (depth < layer_depth) {
-            for (depth + 1..layer_depth) |d| {
-                data[d] = data[depth];
-            }
-        }
-
-        const frame_bytes: usize = if (depth > 0) bytes: {
-            var dif_count: usize = 0;
-            for (data[depth], data[depth - 1]) |new, src| {
-                if (new != src) dif_count += 1;
-            }
-            break :bytes dif_count * 4;
-        } else savestate_size * 4;
-
-        var buf: [17]u8 = undefined;
-        _ = std.fmt.bufPrint(&buf, "Frame {d: >3}:\t{d: >4}\r\n", .{ i + 1, frame_bytes + savestate_head }) catch unreachable;
-        _ = reportfile.write(&buf) catch unreachable;
-        total_bytes += frame_bytes + savestate_head;
-    }
-
-    var buf: [26]u8 = undefined;
-    _ = std.fmt.bufPrint(&buf, "Total: {d: >8}/{d: >8}\r\n", .{ total_bytes, savestate_size * savestate_count * 4 }) catch unreachable;
-    _ = reportfile.write(&buf) catch unreachable;
 }